pyproject.toml
[project]
name = "edgar-rag"
version = "1.0.0"
description = "AI-powered research assistant for SEC EDGAR filings"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "chromadb>=0.5.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.38.0",
    "edgartools>=3.0.0",
    "pick>=2.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
